1064340 - adding temporary puppet executable to run without GC
This is a temporary solution put in place while we wait for the fix to:

https://bugzilla.redhat.com/show_bug.cgi?id=1064340

This patches kafo to use the inlined puppet script that includes a line
to disable garbage collection so the installer doesn't core dump.  I
placed the script in /usr/libexec since the basename of the script must
be 'puppet' and not something different like 'puppet-nogc' which doesn't
work.  I wanted a second script, outside of the user's path to ensure
we don't effect other users on the system calling /usr/bin/puppet